atm_backend/requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==20.1.0
flake8==6.1.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg==3.1.14
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
